Tried making a separate class for a movable object, probably gonna be reworked
